Update Plugins
Appointments+
Black Studio TinyMCE Widget
CoursePress Pro
Events +
Jobs and Experts
MarketPress
New Blog Templates
Shareaholic
Ultimate Branding
Ultimate Member

+++ content/plugins/ultimate-member/admin/core/um-admin-users.php
@@ -35,12 +35,14 @@ class UM_Admin_Users {
 		unset( $actions['edit'] );
 		unset( $actions['delete'] );
 
-		$actions['backend_profile'] = "<a class='' href='" . admin_url('user-edit.php?user_id='. $user_id ) . "'>" . __( 'Edit','ultimatemember' ) . "</a>";
-		$actions['frontend_profile'] = "<a class='' href='" . um_user_profile_url() . "'>" . __( 'Edit in frontend','ultimatemember') . "</a>";
+		$actions['frontend_profile'] = "<a class='' href='" . um_user_profile_url() . "'>" . __( 'Profile','ultimatemember') . "</a>";
 		
 		if ( um_user('submitted') ) {
 			$actions['view_info'] = '<a href="#" data-modal="UM_preview_registration" data-modal-size="smaller" data-dynamic-content="um_admin_review_registration" data-arg1="'.$user_id.'" data-arg2="edit_registration">' . __('Info','ultimatemember') . '</a>';
 		}
+		
+		$actions = apply_filters('um_admin_user_row_actions', $actions, $user_id );
+		
 		return $actions;
 	}
 	
@@ -84,6 +86,16 @@ class UM_Admin_Users {
 		if ( is_admin() && $pagenow=='users.php' && isset($_GET[ 'status' ]) && $_GET[ 'status' ] != '') {
 			
 			$status = urldecode($_GET[ 'status' ]);
+			
+			if ( $status == 'needs-verification') {
+			$query->query_where = str_replace('WHERE 1=1', 
+						"WHERE 1=1 AND {$wpdb->users}.ID IN (
+							 SELECT {$wpdb->usermeta}.user_id FROM $wpdb->usermeta 
+								WHERE {$wpdb->usermeta}.meta_key = '_um_verified' 
+								AND {$wpdb->usermeta}.meta_value = 'pending')", 
+						$query->query_where
+			);
+			} else {
 			$query->query_where = str_replace('WHERE 1=1', 
 						"WHERE 1=1 AND {$wpdb->users}.ID IN (
 							 SELECT {$wpdb->usermeta}.user_id FROM $wpdb->usermeta 
@@ -91,6 +103,7 @@ class UM_Admin_Users {
 								AND {$wpdb->usermeta}.meta_value = '{$status}')", 
 						$query->query_where
 			);
+			}
 
 		}
 
@@ -133,6 +146,8 @@ class UM_Admin_Users {
 			
 			$views[ $k ] = '<a href="'.admin_url('users.php').'?status='.$k.'" ' . $current . '>'. $v . ' <span class="count">('.$ultimatemember->query->count_users_by_status( $k ).')</span></a>';
 		}
+		
+		$views = apply_filters('um_admin_views_users', $views );
 		
 		return $views;
 	}
@@ -304,7 +319,7 @@ class UM_Admin_Users {
 	function manage_users_columns($columns) {
 	
 		$admin = new UM_Admin_Metabox();
-		
+
 		$columns['um_role'] = __('Community Role','ultimatemember') . $admin->_tooltip( __('This is the membership role set by Ultimate Member plugin','ultimatemember') );
 
 		return $columns;
@@ -318,7 +333,9 @@ class UM_Admin_Users {
 
 		if ( $this->custom_role == $column_name ) {
 		
-			delete_option( "um_cache_userdata_{$user_id}" );
+			if ( get_option( "um_cache_userdata_{$user_id}" ) ) {
+				delete_option( "um_cache_userdata_{$user_id}" );
+			}
 			um_fetch_user( $user_id );
 			return um_user('role_name');
